Added three columns to sql table - categories, maturity rating, and average rating.
for category, I needed to add a python list to mysql, which I did through the 'join' method, similar for what I did with authors.

for maturity rating I had to input a boolean value, which I completed by adding an if/else function that would classify trues as 1 and falses as 0

for average rating, i simply converted the average_rating variable into a float and set the default value of average_rating to be 0.0.